+++ src/components/handshakes/exchange/Feed/FeedExchange.scss
@@ -1,97 +1,34 @@
 @import './src/styles/var';
 
-.feed {
-  border-radius: 6px;
+.app {
+  padding-bottom: 150px !important;
 }
-
 .feed-exchange {
-  .name-shop {
-    font-weight: 500;
-    color: #FFFFFF;
-    font-size: 16px;
-    letter-spacing: 0.29px;
+  .coin-item {
+    display: inline-block;
+    width: 50%;
+    color: white;
   }
-  .distance {
-    padding-bottom: 2px;
-    font-weight: 400;
-    font-size: 14px;
-    color: #FFFFFF;
-    img {
-      margin-top: -2px;
-      padding: 0 11px 0 7px;
-    }
-  }
-  .transaction-successful {
-    color: #4CD964;
-    font-weight: 400;
-    font-size: 14px;
-  }
-  .transaction-failed {
-    font-size: 14px;
-    color: #F5A623;
-    font-weight: 400;
-  }
-  .table-ex {
+  .ex-sticky-note {
+    text-align: center;
+    color: white;
+    position: fixed;
+    bottom: 60px;
+    left: 50%;
+    transform: translateX(-50%);
     width: 100%;
-    border-bottom: 1px solid rgba(255,255,255,0.25);
-    .image {
-      vertical-align: middle;
-      display: inline-block;
-      width: 30px;
-      margin-right: 4px;
-      img {
-        width: 100%;
-        display: block;
-        height: auto;
-      }
-    }
-    td, th {
-      border: 0;
-    }
-    thead {
-      border-top: 1px solid rgba(255,255,255,0.25);
-      border-bottom: 1px solid rgba(255,255,255,0.25);
-      th {
-        padding: 5px 0;
-      }
-    }
-    tbody {
-      td {
-        padding-top: 15px;
-      }
-      tr:last-child {
-        td {
-          padding-bottom: 15px;
-        }
-      }
-    }
-    .header-text {
-      font-weight: 500;
-      font-size: 14px;
-      color: #FFFFFF;
-      letter-spacing: 0.25px;
-    }
-    .buy-color {
-      color: #00FFC6;
-    }
-    .sell-color {
-      color: #FF3B3B;
-    }
-    .price-number {
-      font-weight: 900;
-      font-size: 16px;
-      letter-spacing: 0.29px;
-      text-shadow: 0 2px 4px rgba(0,0,0,0.50);
-    }
-    .currency {
-      opacity: 0.8;
-      font-weight: 500;
-      font-size: 14px;
-      color: #FFFFFF;
-      letter-spacing: 0.25px;
-      //line-height: 32px;
+    max-width: 414px;
+    background-color: black;
+    z-index: 2;
+    padding: 10px 0;
+    .btn-become {
+      min-width: 200px;
+      border: 1px solid white;
+      background: black;
+      color: white;
     }
   }
+
   .dialog-shake-detail {
     .modal-custom-body {
       background: #485059;
