+++ src/components/handshakes/exchange/Feed/FeedExchangeBackUp2.scss
@@ -1,0 +1,102 @@
+@import './src/styles/var';
+
+.feed {
+  border-radius: 6px;
+}
+
+.feed-exchange {
+  .name-shop {
+    font-weight: 500;
+    color: #FFFFFF;
+    font-size: 16px;
+    letter-spacing: 0.29px;
+  }
+  .distance {
+    padding-bottom: 2px;
+    font-weight: 400;
+    font-size: 14px;
+    color: #FFFFFF;
+    img {
+      margin-top: -2px;
+      padding: 0 11px 0 7px;
+    }
+  }
+  .transaction-successful {
+    color: #4CD964;
+    font-weight: 400;
+    font-size: 14px;
+  }
+  .transaction-failed {
+    font-size: 14px;
+    color: #F5A623;
+    font-weight: 400;
+  }
+  .table-ex {
+    width: 100%;
+    border-bottom: 1px solid rgba(255,255,255,0.25);
+    .image {
+      vertical-align: middle;
+      display: inline-block;
+      width: 30px;
+      margin-right: 4px;
+      img {
+        width: 100%;
+        display: block;
+        height: auto;
+      }
+    }
+    td, th {
+      border: 0;
+    }
+    thead {
+      border-top: 1px solid rgba(255,255,255,0.25);
+      border-bottom: 1px solid rgba(255,255,255,0.25);
+      th {
+        padding: 5px 0;
+      }
+    }
+    tbody {
+      td {
+        padding-top: 15px;
+      }
+      tr:last-child {
+        td {
+          padding-bottom: 15px;
+        }
+      }
+    }
+    .header-text {
+      font-weight: 500;
+      font-size: 14px;
+      color: #FFFFFF;
+      letter-spacing: 0.25px;
+    }
+    .buy-color {
+      color: #00FFC6;
+    }
+    .sell-color {
+      color: #FF3B3B;
+    }
+    .price-number {
+      font-weight: 900;
+      font-size: 16px;
+      letter-spacing: 0.29px;
+      text-shadow: 0 2px 4px rgba(0,0,0,0.50);
+    }
+    .currency {
+      opacity: 0.8;
+      font-weight: 500;
+      font-size: 14px;
+      color: #FFFFFF;
+      letter-spacing: 0.25px;
+      //line-height: 32px;
+    }
+  }
+  .dialog-shake-detail {
+    .modal-custom-body {
+      background: #485059;
+      border-radius: 4px;
+      padding: 20px !important;
+    }
+  }
+}
